Fixed a number of message handling issues
 - Resolved a race condition in SSLMessageHandler that could cause
out-of-phase errors between client and server. The root cause was the
fact that SSLEngine.unwrap() apparently only decrypts one TLS packet at a
time, which caused issues when multiple packets were backed up in the
network input buffer.

- Removed some unnecessary synchronizations, and tweaked the
handling of interest flags in the SSL message handlers

- Message handlers now return the number of bytes they read/write. Note
that these are counted as bytes of application data. For the plaintext
handler the distinction doesn't matter, but the SSL handler of course
sends strictly more data over the network than it receives from the
application layer.